DRF_projects/myNoticeBoard/requirements.txt
asgiref==3.5.2
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.12.2
Pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3
